feat: create advert title and border
